fix view, collected all chenges

See merge request python/vdrive!20

+++ apps/vdrive/views.py
@@ -1,16 +1,19 @@
 import logging
-import urllib.request
+
 
 from django import forms
 from django.db.transaction import on_commit
+from django.http import HttpResponse
 from django.urls import reverse_lazy
 from django.contrib.auth.mixins import LoginRequiredMixin
+from django.views import View
 from django.views.generic import ListView, FormView
 
 from googleapiclient.discovery import build
 
-from apps.vdrive.tasks import process
-from apps.vdrive.models import VideoProcessing, Processing, Video
+from apps.vdrive.tasks import process, scan_files
+from apps.vdrive.models import VideoProcessing, Processing, Video, VideoScan
+from .scan import scan_gphotos
 from .utils import get_google_credentials
 
 
@@ -27,14 +30,35 @@ class GDriveListForm(forms.Form):
             self.initial[field_id] = False
 
 
+class StartScanView(View):
+    def start_scan(self):
+        if self.request.user.video_scans.filter(status__in=['in_progress', 'waiting']).exists():
+            logger.debug(f'Active scan already exists for user {self.request.user}')
+        else:
+            video_scan = VideoScan.objects.create(user=self.request.user)
+            on_commit(lambda: scan_files.delay(video_scan.id))
+
+    def get(self, request):
+        self.start_scan()
+        return HttpResponse('Ok')
+
+    def post(self, request):
+        self.start_scan()
+        return HttpResponse('Ok')
+
+
 class GDriveListView(LoginRequiredMixin, FormView):
     template_name = 'vdrive/list.html'
     form_class = GDriveListForm
     success_url = reverse_lazy('vdrive:imports_list')
 
     def get_form_kwargs(self):
-        self.get_files_list()
-        self.get_files_list_gphotos()
+        if self.request.user.video_scans.exists():
+            logger.debug(f'Active scan already exists for user {self.request.user}')
+        else:
+            video_scan = VideoScan.objects.create(user=self.request.user)
+            on_commit(lambda: scan_files.delay(video_scan.id))
+
         kwargs = super().get_form_kwargs()
         kwargs['videos'] = self.get_videos()
         return kwargs
@@ -49,69 +73,17 @@ class GDriveListView(LoginRequiredMixin, FormView):
         context['videos'] = self.get_videos()
         return context
 
-    def get_files_list(self):
-        user = self.request.user
-        drive = build('drive', 'v3', credentials=get_google_credentials(user))
-        files_data = drive.files().list(q="mimeType contains 'video/'",
-                                        spaces='drive',
-                                        fields='files(id, name, size, thumbnailLink)'
-                                        ).execute()
-        logger.info(f'Found fies {files_data}')
-
-        for item in files_data["files"]:
-            video = Video.objects.get_or_create(source_id=item['id'],
-                                                user=user,
-                                                source_type=Video.Type.GDRIVE,
-                                                defaults={
-                                                    'name': item['name'],
-                                                    'thumbnail': item.get('thumbnailLink', ''),
-                                                    'size': int(item['size'])
-                                                })
-        return files_data['files']
-
-    def get_files_list_gphotos(self):
-        user = self.request.user
-        library = build('photoslibrary', 'v1', credentials=get_google_credentials(user))
-        results = library.mediaItems().list().execute()
-        logger.info(f'Found gphotos files {results}')
-        items = results.get('mediaItems', [])
-        for item in items:
-            print(item['mimeType'], item['filename'])
-            if not item['mimeType'].startswith('video'):
-                continue
-
-            base_url = item.get('baseUrl')
-            if not base_url:
-                logger.error(f'No base url for video {item}')
-                continue
-
-            download_url = base_url + '=dv'
-            size = 0
-
-            try:
-                with urllib.request.urlopen(download_url) as url_downloader:
-                    size = url_downloader.getheader('content-length')
-            except Exception:
-                logger.error(f'Error fetching size for video {item}')
-
-            Video.objects.get_or_create(source_type=Video.Type.GPHOTOS,
-                                        source_id=item['id'],
-                                        user=self.request.user,
-                                        defaults={
-                                            'thumbnail': base_url + '=w300-h200',
-                                            'name': item.get('filename', 'UNKNOWN'),
-                                            'size': size
-                                        })
-        return items
 
     def form_valid(self, form):
         data = list(form.data)
         videos = [field for field in data if field != 'csrfmiddlewaretoken']
         processing = Processing.objects.create()
+        video_pks = []
         for video_id in videos:
             video_processing = VideoProcessing.objects.create(video_id=video_id, processing=processing)
             video_processing.save()
-            on_commit(lambda: process(video_processing.pk))
+            video_pks.append(video_processing.pk)
+        on_commit(lambda: [process.delay(video_pk) for video_pk in video_pks])
         return super().form_valid(form)
 
 
@@ -121,9 +93,9 @@ class UserListView(LoginRequiredMixin, ListView):
 
 
 class DelListForm(forms.Form):
-    def __init__(self, *args, user=None,  **kwargs):
+    def __init__(self, *args, videos=None, **kwargs):
         super().__init__(*args, **kwargs)
-        for video in user.videos.filter(processings__status='success'):
+        for video in videos:
             field_name = video.name
             field_id = video.id
             self.fields[field_id] = forms.BooleanField(required=False,
@@ -138,23 +110,41 @@ class DeleteListView(LoginRequiredMixin, FormView):
 
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
-        kwargs['user'] = self.request.user
+        kwargs['videos'] = self.get_videos()
         return kwargs
 
-    def get_files_list(self):
-        return
+    def get_videos(self):
+        return self.request.user.videos.filter(processings__status=VideoProcessing.Status.SUCCESS)
+
+    def get_context_data(self, **kwargs):
+        context = super(DeleteListView, self).get_context_data(**kwargs)
+        context['videos'] = self.get_videos()
+        return context
 
     def form_valid(self, form):
         data = list(form.data)
+        print(data)
         videos = [field for field in data if field != 'csrfmiddlewaretoken']
-        drive = build('drive', 'v3', credentials=get_google_credentials(self.request.user))
         for video_id in videos:
-            id = Video.objects.get(id=video_id)
-            id_source = id.source_id
-            drive.files().delete(fileId=id_source).execute()
-            logger.info(f'Deleted file: {id}')
-            id.status = Video.Status.DELETED
-            id.save()
+            video = Video.objects.get(id=video_id)
+            id_source = video.source_id
+
+            if video.source_type == Video.Type.GDRIVE:
+                drive = build('drive', 'v3', credentials=get_google_credentials(self.request.user))
+                video = Video.objects.get(id=video_id)
+                drive.files().delete(fileId=id_source).execute()
+                logger.info(f'Deleted file: {id}')
+                video.status = Video.Status.DELETED
+                video.save()
+            else:
+                library = build('photoslibrary', 'v1', credentials=get_google_credentials(self.request.user))
+                request_body = {
+                    "mediaItemIds": video_id
+                }
+                results = library.batchRemoveMediaItems(body=request_body).execute()
+                logger.info(f'Deleted file: {results}')
+
+                # video.status = Video.Status.DELETED
+                # video.save()
+
         return super().form_valid(form)
-
-
